requirements.txt
python-telegram-bot
yfinance
pandas
pandas_ta
Flask
python-dotenv
feedparser
openai